requirements.txt
pbr>=1.6 # Apache-2.0
oslo.config>=3.14.0 # Apache-2.0
oslo.log>=3.11.0 # Apache-2.0
oslo.utils>=3.16.0 # Apache-2.0
python-ironicclient>=1.6 # Apache-2.0
hpOneView>=4.4.0
futures
